people-search/pyproject.toml
[project]
name = "people-search"
version = "0.1.0"
description = "Search for people in the database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "enrichmcp>=0.4.0",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "langchain-openai>=0.3.23",
    "markdownify>=1.1.0",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.3",
    "uvicorn>=0.34.3",
]
